Update for Evolutek<< 2015
Changes:

- move old code to 2014
- remove unused Makefile and systemd stuff (see Rationale for removing 'make
  upload') below
- store all python code in a single namespace
- move most of the 'robot.py' functionnality to robotctl
- tools for ax12 have moved to robotctl
- robot.py is now part of the 'evolutek' library and can be easily be used by
  services and other projects.

Warning: this change is bleeding edge, mostly untested.

Rationale for removing 'make upload'
------------------------------------

During the Coupde de France 2014 we had several instances of 'wtf is the code
running in the robot'. This is mostly due to the fact that this code was setup
outside the git repository and hot-patched without commiting the changes. This
is bad. Therefore I think we should integrate the robot into our git workflow.
I also think there still will be hot-patching directly on the robot, but if the
code is in a repository we will at least be able to see the changes and commit
them afterwards.